+++ Documentation/RxGRDBDemo/RxGRDBDemo/Player.swift
@@ -1,7 +1,7 @@
 import GRDB
 
 // A player
-struct Player: Codable {
+struct Player: Codable, Equatable {
     var id: Int64?
     var name: String
     var score: Int
@@ -16,8 +16,10 @@ extension Player: FetchableRecord { }
 extension Player: MutablePersistableRecord {
     static let databaseTableName = "player"
     
-    enum Columns: String, ColumnExpression {
-        case id, name, score
+    enum Columns {
+        static let id = Column(CodingKeys.id)
+        static let name = Column(CodingKeys.name)
+        static let score = Column(CodingKeys.score)
     }
     
     mutating func didInsert(with rowID: Int64, for column: String?) {