+++ CHANGELOG.md
@@ -1,6 +1,33 @@
 Release Notes
 =============
 
+## Next Version
+
+### New
+
+- Support for Swift 5, GRDB 4.0, and RxSwift 5.0.
+- Reactive extension on DatabaseRegionObservation:
+    
+    ```swift
+    let players = Player.all()
+    let teams = Team.all()
+    let observation = DatabaseRegionObservation(tracking: players, teams)
+    observation.rx.changes(in: dbQueue)
+        .subscribe(onNext: { db: Database in
+            print("Players or teams have changed.")
+        })
+    ```
+
+### Breaking Changes
+
+- Swift 4.0 and Swift 4.1 are no longer supported.
+- GRDB 3 and RxSwift 4 are no longer supported.
+- iOS 8 is no longer supported. Minimum deployment target is now iOS 9.0.
+- Deprecated APIs are no longer available.
+- `DatabaseWriter.rx.changes` is removed, replaced with `DatabaseRegionObservation.rx.changes`.
+- SQLCipher support is now available under the CocoaPods `RxGRDB/SQLCipher` name.
+
+
 ## 0.13.0
 
 Released November 2, 2018 &bull; [diff](https://github.com/RxSwiftCommunity/RxGRDB/compare/v0.12.1...v0.13.0)